refac: make domain folder for entities
