+++ src/main/java/org/example/kkumdoriland/dream/domain/DailyHistory.java
@@ -1,4 +1,4 @@
-package org.example.kkumdoriland.dream;
+package org.example.kkumdoriland.dream.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -7,23 +7,19 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import java.time.LocalDateTime;
 import org.example.kkumdoriland.common.domain.BaseEntity;
-import org.example.kkumdoriland.member.domain.Member;
 
 @Entity
-public class Dream extends BaseEntity {
+public class DailyHistory extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private String title;
-
     @Column(columnDefinition = "text")
-    private String description;
-    private LocalDateTime dueDate;
+    private String contentText;
+    private int stepSize;
 
     @ManyToOne
-    @JoinColumn(name="ownerId")
-    private Member user;
+    @JoinColumn(name="mileStoneId")
+    private MileStone mileStone;
 }
