+++ src/main/java/org/example/kkumdoriland/dream/domain/Dream.java
@@ -1,4 +1,4 @@
-package org.example.kkumdoriland.dream;
+package org.example.kkumdoriland.dream.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -7,19 +7,23 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import java.time.LocalDateTime;
 import org.example.kkumdoriland.common.domain.BaseEntity;
+import org.example.kkumdoriland.member.domain.Member;
 
 @Entity
-public class DailyHistory extends BaseEntity {
+public class Dream extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    private String title;
+
     @Column(columnDefinition = "text")
-    private String contentText;
-    private int stepSize;
+    private String description;
+    private LocalDateTime dueDate;
 
     @ManyToOne
-    @JoinColumn(name="mileStoneId")
-    private MileStone mileStone;
+    @JoinColumn(name="ownerId")
+    private Member user;
 }
