+++ src/main/java/org/example/kkumdoriland/dream/domain/MileStone.java
@@ -1,4 +1,4 @@
-package org.example.kkumdoriland.dream;
+package org.example.kkumdoriland.dream.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
